Start restructuring Graph traversal into temporal and non-temporal
